+++ src/game/Object/Creature.cpp
@@ -1624,193 +1624,6 @@ void Creature::SaveToDB(uint32 mapid, uint8 spawnMask)
 }
 
 /**
- * @brief Selects the creature level and recalculates level-dependent stats.
- *
- * @param forcedLevel Optional forced level override.
- */
-void Creature::SelectLevel(uint32 forcedLevel /*= USE_DEFAULT_DATABASE_LEVEL*/)
-{
-    CreatureInfo const* cinfo = GetCreatureInfo();
-    if (!cinfo)
-    {
-        return;
-    }
-
-    uint32 rank = IsPet() ? 0 : cinfo->Rank;                // TODO :: IsPet probably not needed here
-
-    // level
-    uint32 level = forcedLevel;
-    uint32 const minlevel = cinfo->MinLevel;
-    uint32 const maxlevel = cinfo->MaxLevel;
-
-    if (level == USE_DEFAULT_DATABASE_LEVEL)
-    {
-        level = minlevel == maxlevel ? minlevel : urand(minlevel, maxlevel);
-    }
-
-    SetLevel(level);
-
-    //////////////////////////////////////////////////////////////////////////
-    // Calculate level dependent stats
-    //////////////////////////////////////////////////////////////////////////
-
-    uint32 health;
-    uint32 mana;
-
-    if (CreatureClassLvlStats const* cCLS = sObjectMgr.GetCreatureClassLvlStats(level, cinfo->UnitClass, cinfo->Expansion))
-    {
-        // Use Creature Stats to calculate stat values
-
-        // health
-        health = cCLS->BaseHealth * cinfo->HealthMultiplier;
-
-        // mana
-        mana = cCLS->BaseMana * cinfo->PowerMultiplier;
-    }
-    else
-    {
-        if (forcedLevel == USE_DEFAULT_DATABASE_LEVEL || (forcedLevel >= minlevel && forcedLevel <= maxlevel))
-        {
-            // Use old style to calculate stat values
-            float rellevel = maxlevel == minlevel ? 0 : (float(level - minlevel)) / (maxlevel - minlevel);
-
-            // health
-            uint32 minhealth = std::min(cinfo->MaxLevelHealth, cinfo->MinLevelHealth);
-            uint32 maxhealth = std::max(cinfo->MaxLevelHealth, cinfo->MinLevelHealth);
-            health = uint32(minhealth + uint32(rellevel * (maxhealth - minhealth)));
-
-            // mana
-            uint32 minmana = std::min(cinfo->MaxLevelMana, cinfo->MinLevelMana);
-            uint32 maxmana = std::max(cinfo->MaxLevelMana, cinfo->MinLevelMana);
-            mana = minmana + uint32(rellevel * (maxmana - minmana));
-        }
-        else
-        {
-            sLog.outError("Creature::SelectLevel> Error trying to set level(%u) for creature %s without enough data to do it!", level, GetGuidStr().c_str());
-            // probably wrong
-            health = (cinfo->MaxLevelHealth / cinfo->MaxLevel) * level;
-            mana = (cinfo->MaxLevelMana / cinfo->MaxLevel) * level;
-        }
-    }
-
-    health *= _GetHealthMod(rank); // Apply custom config setting
-    if (health < 1)
-    {
-        health = 1;
-    }
-
-    //////////////////////////////////////////////////////////////////////////
-    // Set values
-    //////////////////////////////////////////////////////////////////////////
-
-    // health
-    SetCreateHealth(health);
-    SetMaxHealth(health);
-    SetHealth(health);
-
-    SetModifierValue(UNIT_MOD_HEALTH, BASE_VALUE, float(health));
-
-    // all power types. WARNING!!! pay great attention to MAX_POWER value
-    // because you can easily overwrite the UNIT_FIELD_LEVEL
-    for (int i = POWER_MANA; i < MAX_POWERS; ++i)
-    {
-        uint32 maxValue = 0;
-
-        switch (i)
-        {
-            case POWER_MANA:        maxValue = mana; break;
-            case POWER_RAGE:        maxValue = 0; break;
-            case POWER_FOCUS:       maxValue = POWER_FOCUS_DEFAULT; break;
-            case POWER_ENERGY:      maxValue = POWER_ENERGY_DEFAULT * cinfo->PowerMultiplier; break;
-            case POWER_HAPPINESS:   maxValue = POWER_HAPPINESS_DEFAULT; break;
-            default:                break; // make compilers happy
-        }
-
-        uint32 value = maxValue;
-
-        // For non regenerating powers set 0
-        if ((i == POWER_ENERGY || i == POWER_MANA) && !IsRegeneratingPower())
-        {
-            value = 0;
-        }
-
-        // Mana requires an extra field to be set
-        if (i == POWER_MANA)
-        {
-            SetCreateMana(value);
-        }
-
-        SetMaxPower(Powers(i), maxValue);
-        SetPower(Powers(i), value);
-        SetModifierValue(UnitMods(UNIT_MOD_POWER_START + i), BASE_VALUE, float(value));
-    }
-
-    // damage
-    float damagemod = _GetDamageMod(rank);
-
-    SetBaseWeaponDamage(BASE_ATTACK, MINDAMAGE, cinfo->MinMeleeDmg * damagemod);
-    SetBaseWeaponDamage(BASE_ATTACK, MAXDAMAGE, cinfo->MaxMeleeDmg * damagemod);
-
-    SetBaseWeaponDamage(OFF_ATTACK, MINDAMAGE, cinfo->MinMeleeDmg * damagemod);
-    SetBaseWeaponDamage(OFF_ATTACK, MAXDAMAGE, cinfo->MaxMeleeDmg * damagemod);
-
-    SetFloatValue(UNIT_FIELD_MINRANGEDDAMAGE, cinfo->MinRangedDmg * damagemod);
-    SetFloatValue(UNIT_FIELD_MAXRANGEDDAMAGE, cinfo->MaxRangedDmg * damagemod);
-
-    SetModifierValue(UNIT_MOD_ATTACK_POWER, BASE_VALUE, cinfo->MeleeAttackPower * damagemod);
-}
-
-/**
- * @brief Gets the configured health multiplier for a creature rank.
- *
- * @param Rank The creature elite rank.
- * @return The configured health rate multiplier.
- */
-float Creature::_GetHealthMod(int32 Rank)
-{
-    switch (Rank)                                           // define rates for each elite rank
-    {
-        case CREATURE_ELITE_NORMAL:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_NORMAL_HP);
-        case CREATURE_ELITE_ELITE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_ELITE_HP);
-        case CREATURE_ELITE_RAREELITE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_RAREELITE_HP);
-        case CREATURE_ELITE_WORLDBOSS:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_WORLDBOSS_HP);
-        case CREATURE_ELITE_RARE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_RARE_HP);
-        default:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_ELITE_HP);
-    }
-}
-
-/**
- * @brief Gets the configured damage multiplier for a creature rank.
- *
- * @param Rank The creature elite rank.
- * @return The configured damage rate multiplier.
- */
-float Creature::_GetDamageMod(int32 Rank)
-{
-    switch (Rank)                                           // define rates for each elite rank
-    {
-        case CREATURE_ELITE_NORMAL:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_NORMAL_DAMAGE);
-        case CREATURE_ELITE_ELITE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_ELITE_DAMAGE);
-        case CREATURE_ELITE_RAREELITE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_RAREELITE_DAMAGE);
-        case CREATURE_ELITE_WORLDBOSS:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_WORLDBOSS_DAMAGE);
-        case CREATURE_ELITE_RARE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_RARE_DAMAGE);
-        default:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_ELITE_DAMAGE);
-    }
-}
-
-/**
  * @brief Lowers the remaining player damage requirement for loot eligibility.
  *
  * @param unDamage The damage amount to subtract.
@@ -1820,31 +1633,6 @@ void Creature::LowerPlayerDamageReq(uint32 unDamage)
     if (m_PlayerDamageReq)
     {
         m_PlayerDamageReq > unDamage ? m_PlayerDamageReq -= unDamage : m_PlayerDamageReq = 0;
-    }
-}
-
-/**
- * @brief Gets the spell damage modifier for a creature elite rank.
- *
- * @param Rank The creature elite rank.
- * @return The configured spell damage multiplier.
- */
-float Creature::_GetSpellDamageMod(int32 Rank)
-{
-    switch (Rank)                                           // define rates for each elite rank
-    {
-        case CREATURE_ELITE_NORMAL:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_NORMAL_SPELLDAMAGE);
-        case CREATURE_ELITE_ELITE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_ELITE_SPELLDAMAGE);
-        case CREATURE_ELITE_RAREELITE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_RAREELITE_SPELLDAMAGE);
-        case CREATURE_ELITE_WORLDBOSS:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_WORLDBOSS_SPELLDAMAGE);
-        case CREATURE_ELITE_RARE:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_RARE_SPELLDAMAGE);
-        default:
-            return sWorld.getConfig(CONFIG_FLOAT_RATE_CREATURE_ELITE_ELITE_SPELLDAMAGE);
     }
 }
 
@@ -3103,96 +2891,6 @@ Unit* Creature::SelectAttackingTarget(AttackingTarget target, uint32 position, S
 }
 
 /**
- * @brief Stores an absolute cooldown end time for a creature spell.
- *
- * @param spell_id The spell identifier.
- * @param end_time The cooldown end time.
- */
-void Creature::_AddCreatureSpellCooldown(uint32 spell_id, time_t end_time)
-{
-    m_CreatureSpellCooldowns[spell_id] = end_time;
-}
-
-/**
- * @brief Stores the application time for a spell category cooldown.
- *
- * @param category The spell category.
- * @param apply_time The time when the category cooldown started.
- */
-void Creature::_AddCreatureCategoryCooldown(uint32 category, time_t apply_time)
-{
-    m_CreatureCategoryCooldowns[category] = apply_time;
-}
-
-/**
- * @brief Adds cooldown tracking for a creature spell and its category.
- *
- * @param spellid The spell identifier.
- */
-void Creature::AddCreatureSpellCooldown(uint32 spellid)
-{
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spellid);
-    if (!spellInfo)
-    {
-        return;
-    }
-
-    uint32 cooldown = GetSpellRecoveryTime(spellInfo);
-    if (cooldown)
-    {
-        _AddCreatureSpellCooldown(spellid, time(NULL) + cooldown / IN_MILLISECONDS);
-    }
-
-    if (spellInfo->Category)
-    {
-        _AddCreatureCategoryCooldown(spellInfo->Category, time(NULL));
-    }
-}
-
-/**
- * @brief Checks whether a spell category cooldown is still active.
- *
- * @param spell_id The spell identifier.
- * @return true if the category cooldown is active; otherwise, false.
- */
-bool Creature::HasCategoryCooldown(uint32 spell_id) const
-{
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell_id);
-    if (!spellInfo)
-    {
-        return false;
-    }
-
-    CreatureSpellCooldowns::const_iterator itr = m_CreatureCategoryCooldowns.find(spellInfo->Category);
-    return (itr != m_CreatureCategoryCooldowns.end() && time_t(itr->second + (spellInfo->CategoryRecoveryTime / IN_MILLISECONDS)) > time(NULL));
-}
-
-/**
- * @brief Gets the remaining cooldown delay for a creature spell.
- *
- * @param spellId The spell identifier.
- * @return Remaining cooldown in seconds.
- */
-uint32 Creature::GetCreatureSpellCooldownDelay(uint32 spellId) const
-{
-    CreatureSpellCooldowns::const_iterator itr = m_CreatureSpellCooldowns.find(spellId);
-    time_t t = time(NULL);
-    return uint32(itr != m_CreatureSpellCooldowns.end() && itr->second > t ? itr->second - t : 0);
-}
-
-/**
- * @brief Checks whether a spell or its category is currently on cooldown.
- *
- * @param spell_id The spell identifier.
- * @return true if a cooldown is active; otherwise, false.
- */
-bool Creature::HasSpellCooldown(uint32 spell_id) const
-{
-    CreatureSpellCooldowns::const_iterator itr = m_CreatureSpellCooldowns.find(spell_id);
-    return (itr != m_CreatureSpellCooldowns.end() && itr->second > time(NULL)) || HasCategoryCooldown(spell_id);
-}
-
-/**
  * @brief Checks whether the creature is currently evading back home.
  *
  * @return true if the home movement generator is active; otherwise, false.
@@ -3378,151 +3076,6 @@ uint32 Creature::GetScriptId() const
 {
     // scripts bound to DB guid have priority over ones bound to creature entry
     return sScriptMgr.GetBoundScriptId(SCRIPTED_UNIT, -int32(GetGUIDLow())) ? sScriptMgr.GetBoundScriptId(SCRIPTED_UNIT, -int32(GetGUIDLow())) : sScriptMgr.GetBoundScriptId(SCRIPTED_UNIT, GetEntry());
-}
-
-/**
- * @brief Gets vendor items directly assigned to this creature entry.
- *
- * @return The vendor item list, or null if none exists.
- */
-VendorItemData const* Creature::GetVendorItems() const
-{
-    return sObjectMgr.GetNpcVendorItemList(GetEntry());
-}
-
-/**
- * @brief Gets vendor items from this creature's vendor template.
- *
- * @return The vendor template item list, or null if none exists.
- */
-VendorItemData const* Creature::GetVendorTemplateItems() const
-{
-    uint32 VendorTemplateId = GetCreatureInfo()->VendorTemplateId;
-    return VendorTemplateId ? sObjectMgr.GetNpcVendorTemplateItemList(VendorTemplateId) : NULL;
-}
-
-/**
- * @brief Gets the current available stock count for a vendor item.
- *
- * @param vItem The vendor item definition.
- * @return The currently available count.
- */
-uint32 Creature::GetVendorItemCurrentCount(VendorItem const* vItem)
-{
-    if (!vItem->maxcount)
-    {
-        return vItem->maxcount;
-    }
-
-    VendorItemCounts::iterator itr = m_vendorItemCounts.begin();
-    for (; itr != m_vendorItemCounts.end(); ++itr)
-    {
-        if (itr->itemId == vItem->item)
-        {
-            break;
-        }
-    }
-
-    if (itr == m_vendorItemCounts.end())
-    {
-        return vItem->maxcount;
-    }
-
-    VendorItemCount* vCount = &*itr;
-
-    time_t ptime = time(NULL);
-
-    if (vCount->lastIncrementTime + vItem->incrtime <= ptime)
-    {
-        ItemPrototype const* pProto = ObjectMgr::GetItemPrototype(vItem->item);
-
-        uint32 diff = uint32((ptime - vCount->lastIncrementTime) / vItem->incrtime);
-        if ((vCount->count + diff * pProto->BuyCount) >= vItem->maxcount)
-        {
-            m_vendorItemCounts.erase(itr);
-            return vItem->maxcount;
-        }
-
-        vCount->count += diff * pProto->BuyCount;
-        vCount->lastIncrementTime = ptime;
-    }
-
-    return vCount->count;
-}
-
-/**
- * @brief Updates and consumes stock count for a limited vendor item.
- *
- * @param vItem The vendor item definition.
- * @param used_count The amount being purchased.
- * @return The remaining count after the update.
- */
-uint32 Creature::UpdateVendorItemCurrentCount(VendorItem const* vItem, uint32 used_count)
-{
-    if (!vItem->maxcount)
-    {
-        return 0;
-    }
-
-    VendorItemCounts::iterator itr = m_vendorItemCounts.begin();
-    for (; itr != m_vendorItemCounts.end(); ++itr)
-    {
-        if (itr->itemId == vItem->item)
-        {
-            break;
-        }
-    }
-
-    if (itr == m_vendorItemCounts.end())
-    {
-        uint32 new_count = vItem->maxcount > used_count ? vItem->maxcount - used_count : 0;
-        m_vendorItemCounts.push_back(VendorItemCount(vItem->item, new_count));
-        return new_count;
-    }
-
-    VendorItemCount* vCount = &*itr;
-
-    time_t ptime = time(NULL);
-
-    if (vCount->lastIncrementTime + vItem->incrtime <= ptime)
-    {
-        ItemPrototype const* pProto = ObjectMgr::GetItemPrototype(vItem->item);
-
-        uint32 diff = uint32((ptime - vCount->lastIncrementTime) / vItem->incrtime);
-        if ((vCount->count + diff * pProto->BuyCount) < vItem->maxcount)
-        {
-            vCount->count += diff * pProto->BuyCount;
-        }
-        else
-        {
-            vCount->count = vItem->maxcount;
-        }
-    }
-
-    vCount->count = vCount->count > used_count ? vCount->count - used_count : 0;
-    vCount->lastIncrementTime = ptime;
-    return vCount->count;
-}
-
-/**
- * @brief Gets trainer spells from the trainer template bound to this creature.
- *
- * @return The trainer template spell list, or null if none exists.
- */
-TrainerSpellData const* Creature::GetTrainerTemplateSpells() const
-{
-    uint32 TrainerTemplateId = GetCreatureInfo()->TrainerTemplateId;
-    return TrainerTemplateId ? sObjectMgr.GetNpcTrainerTemplateSpells(TrainerTemplateId) : NULL;
-}
-
-/**
- * @brief Gets trainer spells directly assigned to this creature entry.
- *
- * @return The trainer spell list, or null if none exists.
- */
-TrainerSpellData const* Creature::GetTrainerSpells() const
-{
-    return sObjectMgr.GetNpcTrainerSpells(GetEntry());
 }
 
 // overwrite WorldObject function for proper name localization
@@ -3795,192 +3348,4 @@ void Creature::SetVirtualItemRaw(VirtualItemSlot slot, uint32 display_id, uint32
     SetUInt32Value(UNIT_VIRTUAL_ITEM_SLOT_DISPLAY + slot, display_id);
     SetUInt32Value(UNIT_VIRTUAL_ITEM_INFO + (slot * 2) + 0, info0);
     SetUInt32Value(UNIT_VIRTUAL_ITEM_INFO + (slot * 2) + 1, info1);
-}
-
-/**
- * @brief Enables or disables walk mode for the creature.
- *
- * @param enable true to walk; false to run.
- * @param asDefault true to also update the default running state.
- */
-void Creature::SetWalk(bool enable, bool asDefault)
-{
-    if (asDefault)
-    {
-        if (enable)
-        {
-            clearUnitState(UNIT_STAT_RUNNING);
-        }
-        else
-        {
-            addUnitState(UNIT_STAT_RUNNING);
-        }
-    }
-
-    // Nothing changed?
-    if (enable == m_movementInfo.HasMovementFlag(MOVEFLAG_WALK_MODE))
-    {
-        return;
-    }
-
-    if (enable)
-    {
-        m_movementInfo.AddMovementFlag(MOVEFLAG_WALK_MODE);
-    }
-    else
-    {
-        m_movementInfo.RemoveMovementFlag(MOVEFLAG_WALK_MODE);
-    }
-
-    WorldPacket data(enable ? SMSG_SPLINE_MOVE_SET_WALK_MODE : SMSG_SPLINE_MOVE_SET_RUN_MODE, 9);
-    data << GetPackGUID();
-    SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Enables or disables levitation movement flags.
- *
- * @param enable true to levitate; false to clear the flag.
- */
-void Creature::SetLevitate(bool enable)
-{
-    if (enable)
-    {
-        m_movementInfo.AddMovementFlag(MOVEFLAG_LEVITATING);
-    }
-    else
-    {
-        m_movementInfo.RemoveMovementFlag(MOVEFLAG_LEVITATING);
-    }
-
-    // TODO: there should be analogic opcode for 2.43
-    // WorldPacket data(enable ? SMSG_SPLINE_MOVE_GRAVITY_DISABLE : SMSG_SPLINE_MOVE_GRAVITY_ENABLE, 9);
-    // data << GetPackGUID();
-    // SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Enables or disables swim movement flags and broadcasts the change.
- *
- * @param enable true to swim; false to stop swimming.
- */
-void Creature::SetSwim(bool enable)
-{
-    if (enable)
-    {
-        m_movementInfo.AddMovementFlag(MOVEFLAG_SWIMMING);
-    }
-    else
-    {
-        m_movementInfo.RemoveMovementFlag(MOVEFLAG_SWIMMING);
-    }
-
-    WorldPacket data(enable ? SMSG_SPLINE_MOVE_START_SWIM : SMSG_SPLINE_MOVE_STOP_SWIM);
-    data << GetPackGUID();
-    SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Placeholder for enabling or disabling flight.
- *
- * @param enable Unused flight toggle.
- */
-void Creature::SetCanFly(bool enable)
-{
-    if (enable)
-    {
-        m_movementInfo.AddMovementFlag(MOVEFLAG_CAN_FLY);
-    }
-    else
-    {
-        m_movementInfo.RemoveMovementFlag(MOVEFLAG_CAN_FLY);
-    }
-
-    WorldPacket data(enable ? SMSG_SPLINE_MOVE_SET_FLYING : SMSG_SPLINE_MOVE_UNSET_FLYING, 9);
-    data << GetPackGUID();
-    SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Enables or disables feather-fall movement behavior.
- *
- * @param enable true to enable feather fall; false to restore normal falling.
- */
-void Creature::SetFeatherFall(bool enable)
-{
-    if (enable)
-    {
-        m_movementInfo.AddMovementFlag(MOVEFLAG_SAFE_FALL);
-    }
-    else
-    {
-        m_movementInfo.RemoveMovementFlag(MOVEFLAG_SAFE_FALL);
-    }
-
-    WorldPacket data(enable ? SMSG_SPLINE_MOVE_FEATHER_FALL : SMSG_SPLINE_MOVE_NORMAL_FALL);
-    data << GetPackGUID();
-    SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Enables or disables hover movement behavior.
- *
- * @param enable true to hover; false to unset hover.
- */
-void Creature::SetHover(bool enable)
-{
-    if (enable)
-    {
-        m_movementInfo.AddMovementFlag(MOVEFLAG_HOVER);
-    }
-    else
-    {
-        m_movementInfo.RemoveMovementFlag(MOVEFLAG_HOVER);
-    }
-
-    WorldPacket data(enable ? SMSG_SPLINE_MOVE_SET_HOVER : SMSG_SPLINE_MOVE_UNSET_HOVER, 9);
-    data << GetPackGUID();
-    SendMessageToSet(&data, false);
-}
-
-/**
- * @brief Enables or disables root movement behavior.
- *
- * @param enable true to root; false to unroot.
- */
-void Creature::SetRoot(bool enable)
-{
-    if (enable)
-    {
-        m_movementInfo.AddMovementFlag(MOVEFLAG_ROOT);
-    }
-    else
-    {
-        m_movementInfo.RemoveMovementFlag(MOVEFLAG_ROOT);
-    }
-
-    WorldPacket data(enable ? SMSG_SPLINE_MOVE_ROOT : SMSG_SPLINE_MOVE_UNROOT, 9);
-    data << GetPackGUID();
-    SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Enables or disables water-walking behavior.
- *
- * @param enable true to water walk; false to restore land walking.
- */
-void Creature::SetWaterWalk(bool enable)
-{
-    if (enable)
-    {
-        m_movementInfo.AddMovementFlag(MOVEFLAG_WATERWALKING);
-    }
-    else
-    {
-        m_movementInfo.RemoveMovementFlag(MOVEFLAG_WATERWALKING);
-    }
-
-    WorldPacket data(enable ? SMSG_SPLINE_MOVE_WATER_WALK : SMSG_SPLINE_MOVE_LAND_WALK, 9);
-    data << GetPackGUID();
-    SendMessageToSet(&data, true);
 }
